intelligent-document-processing-agent/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.7
pydantic>=2.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
PyYAML>=6.0